implement provider all screen
