--- lib/Controller/HomeController.dart
@@ -1,103 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:image_picker/image_picker.dart';
-import 'package:google_ml_kit/google_ml_kit.dart';
-import 'package:flutter/services.dart';
-
-class HomeController {
-  final ImagePicker _picker = ImagePicker();
-
-  void displayExtractedText(BuildContext context, String extractedText) {
-    TextEditingController textController = TextEditingController(text: extractedText);
-
-    showDialog(
-      context: context,
-      barrierDismissible: false, 
-      builder: (BuildContext context) {
-        return AlertDialog(
-          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
-          title: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text("Extracted Text", style: TextStyle(fontWeight: FontWeight.bold)),
-            ],
-          ),
-          content: Container(
-            width: double.maxFinite,
-            padding: EdgeInsets.all(8),
-            decoration: BoxDecoration(
-              color: Colors.grey[100],
-              borderRadius: BorderRadius.circular(8),
-            ),
-            child: TextField(
-              controller: textController,
-              maxLines: null,
-              readOnly: true,
-              style: TextStyle(color: Colors.black87),
-              decoration: InputDecoration(
-                border: InputBorder.none,
-                hintText: "Extracted text will appear here",
-              ),
-            ),
-          ),
-          actions: [
-            TextButton(
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-              child: Text("Close", style: TextStyle(color: Colors.red)),
-            ),
-            ElevatedButton.icon(
-              onPressed: () {
-                Clipboard.setData(ClipboardData(text: extractedText));
-                ScaffoldMessenger.of(context).showSnackBar(
-                  SnackBar(content: Text("All text copied to clipboard!")),
-                );
-                Navigator.of(context).pop();
-              },
-              icon: Icon(Icons.copy, size: 18),
-              label: Text("Copy All"),
-              style: ElevatedButton.styleFrom(
-                foregroundColor: Colors.white, backgroundColor: Colors.blueAccent,
-                shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
-              ),
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  Future<void> selectImage(BuildContext context) async {
-    final source = await showDialog<ImageSource>(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
-          title: Text("Select Image Source", style: TextStyle(fontWeight: FontWeight.bold)),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.of(context).pop(ImageSource.camera),
-              child: Text("Camera", style: TextStyle(color: Colors.blue)),
-            ),
-            TextButton(
-              onPressed: () => Navigator.of(context).pop(ImageSource.gallery),
-              child: Text("Gallery", style: TextStyle(color: Colors.green)),
-            ),
-          ],
-        );
-      },
-    );
-
-    if (source == null) return;
-
-    final pickedImage = await _picker.pickImage(source: source);
-    if (pickedImage == null) return;
-
-    final inputImage = InputImage.fromFilePath(pickedImage.path);
-    final textRecognizer = GoogleMlKit.vision.textRecognizer();
-    final recognizedText = await textRecognizer.processImage(inputImage);
-    await textRecognizer.close();
-
-    displayExtractedText(context, recognizedText.text);
-  }
-}
