+++ lib/Screen/HomeScreen/Manager/HomeScreenManager.dart
@@ -1,10 +1,9 @@
 import 'package:google_mlkit_document_scanner/google_mlkit_document_scanner.dart';
 import 'package:image_picker/image_picker.dart';
 
-
-class Controller{
-   late DocumentScanner documentScanner;
-   late DocumentScannerOptions options;
+class homeScreenManger {
+  // late DocumentScanner documentScanner;
+  // late DocumentScannerOptions options;
   bool isScanning = false;
   DocumentScanningResult? scanResult;
   List<String> scannedImages = [];
@@ -12,6 +11,6 @@ class Controller{
   final ImagePicker picker = ImagePicker();
   List<String> pdfHistory = [];
   String? generatedPdfPath;
-
-  
+  String? userEmail;
+ 
 }
