+++ lib/Screen/LoginScreen/LoginScreen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:flutterpracticeversion22/Screen/HomeScreen/HomeScreen.dart';
+import 'package:flutterpracticeversion22/Screen/HomeScreen/ViewScreen/HomeScreen.dart';
 import 'package:flutterpracticeversion22/Screen/SigninScreen/SigninScreen.dart';
 import 'package:flutterpracticeversion22/Screen/SignupScreen/SignupScreen.dart';
