+++ lib/Screen/ProfileScreen/Manager/ProfileManager.dart
@@ -1,0 +1,9 @@
+import 'package:firebase_auth/firebase_auth.dart';
+
+class ProfileManager {
+   User? user;
+  String? displayName;
+  String? email;
+  String? photoURL; 
+
+}
