+++ lib/Screen/WordToPdfScreen/Manager/WordToPdfManager.dart
@@ -1,0 +1,8 @@
+import 'dart:io';
+
+class WordToPdfManager {
+   File? wordFile;
+  File? pdfFile;
+  bool isLoading = false; 
+
+}
