+++ lib/Screen/pdftowordScreen/Manager/PdfToWordManger.dart
@@ -1,0 +1,8 @@
+import 'dart:io';
+
+class PdftoWordManger {
+   File? pdfFile;
+  File? wordFile;
+  bool isLoading = false; 
+
+}
